+++ click_extended/decorators/convert/__init__.py
@@ -1,5 +1,6 @@
 """Initialization file for the `click_extended.decorators.convert` module."""
 
+from click_extended.decorators.convert.convert_area import convert_area
 from click_extended.decorators.convert.convert_bits import convert_bits
 from click_extended.decorators.convert.convert_distance import convert_distance
 from click_extended.decorators.convert.convert_speed import convert_speed
@@ -11,6 +12,7 @@ from click_extended.decorators.convert.convert_volume import convert_volume
 from click_extended.decorators.convert.convert_weight import convert_weight
 
 __all__ = [
+    "convert_area",
     "convert_bits",
     "convert_distance",
     "convert_time",